changed the Float data type in the constructors of all the items to doubles. added some testing code to Main class. completed functionality for the order queue in Order class. also completed customer class.

+++ src/CPU.java
@@ -1,5 +1,5 @@
 public class CPU extends Item{
-    public CPU (String name, Float price, int wattage){ //constructor
+    public CPU (String name, double price, int wattage){ //constructor
         this.setItemName(name);
         this.setItemCost(price);
         this.setWattage(wattage);
