+++ src/PowerSupply.java
@@ -1,5 +1,5 @@
 public class PowerSupply extends Item{
-    public PowerSupply (String name, Float price, int wattage){ //constructor
+    public PowerSupply (String name, double price, int wattage){ //constructor
         this.setItemName(name);
         this.setItemCost(price);
         this.setWattage(wattage);
